requirements.txt
Flask==2.3.3
requests==2.31.0
pandas==2.0.3
smartapi-python==1.2.0
